Refactor imports to make lightweights docker (catboost and sklearn are not included)
