+++ app/model.py
@@ -3,19 +3,13 @@ import pandas as pd
 from sklearn.model_selection import train_test_split
 from sklearn.metrics import accuracy_score
 from train_preprocess import predict_age, cut_outlier
-import onnxruntime
-
-# Uncomment this line if want to use\train catboost instead of onnx model
-#from catboost import Pool, CatBoostClassifier
+from catboost import Pool, CatBoostClassifier
 
 
-def predict(data, onnx=True):
-    if onnx:
-        return load_onnx_model(data)
-    else:
-        model = load_model()
-        model_result = model.predict(data)
-        return model_result
+def predict(data):
+    model = load_model()
+    model_result = model.predict(data)
+    return model_result
 
 
 def xy_split(data):
@@ -97,33 +91,6 @@ def load_model():
     return model
 
 
-def load_onnx_model(df):
-    values = prepare_onnx_vals(df)
-    model = get_onnx_model()
-
-    onnx_pred_info(model, values)
-    result = onnx_inference(model, values)
-    return result[0]  # First value since we need only one
-
-
-def get_onnx_model():
-    return onnxruntime.InferenceSession('models/titanicboost.onnx')
-
-
-def prepare_onnx_vals(df):
-    return df.values.astype('float32')
-
-
-def onnx_inference(model, values):
-    label = model.run(['label'], {'features': values})
-    return label
-
-
-def onnx_pred_info(model, vals):
-    predictions = model.run(['probabilities'], {'features': vals})
-    print('all predictions', predictions)
-
-
 if __name__ == '__main__':
-    # create_model()
+    create_model()
     save_onnx_model()
