+++ app/train_preprocess.py
@@ -1,64 +1,7 @@
 import pandas as pd
 from data_encoder import create_encoder, encode_data
 from predict_age_model import age_predictor
-
-
-def get_title(df):
-    df['Title'] = df['Name'].str.extract(' ([A-Za-z]+)\.', expand=False)
-    df['Title'] = df['Title'].replace(['Dr', 'Col', 'Sir', 'Major', 'Master'], 'Mr')
-    df['Title'] = df['Title'].replace(['Capt', 'Don', 'Jonkheer'], 'Rev')
-    df['Title'] = df['Title'].replace(['Ms', 'Lady', 'Mlle', 'Countess', 'Mme', 'Dona'], 'Rev_fem')
-    return df
-
-
-def get_family(df):
-    df['Family_size'] = df['Parch']+df['SibSp'] + 1
-    return df
-
-
-def get_deck(df):
-    df['Cabin'].fillna('Z', inplace=True)
-    df['Deck'] = df['Cabin'].str.extract(r'([A-Za-z]+)')
-    return df
-
-
-def drop_useless(df):
-    df = df.drop(['PassengerId', 'Name', 'Cabin', 'Ticket'], axis=1)
-    return df
-
-
-def get_age_buckets(df):
-    # labels are 'Child', 'Young Adult', 'Adult', 'Senior'
-    df['AgeBucket'] = pd.cut(df['Age'], bins=[0, 18, 30, 50, 80], labels=[1, 2, 3, 4])
-    # print('empty buckets: ', df[df['AgeBucket'].isna()])
-    df['AgeBucket'] = df['AgeBucket'].astype('int16')
-    return df
-
-
-def get_fare_per_person(df):
-    df['PersonFare'] = df['Fare'] / df['Family_size']
-    return df
-
-
-def fill_na(df):
-    df['Fare'].fillna(df['Fare'].mean(), inplace=True)
-    df['Embarked'].fillna(df['Embarked'].mode()[0], inplace=True)
-    return df
-
-
-def apply_all(data):
-    df = data.copy()
-    for func in [fill_na, get_title, get_family, get_fare_per_person, get_deck,
-                 drop_useless]:  # , predict_age, get_age_buckets, change_dtypes]:
-        df = func(df)
-    df = reassemble_order(df)
-    return df
-
-def reassemble_order(data):
-    # Make sure columns a properly organized since catboost models are sensitive to the column order
-    new_data = data[['Pclass', 'Sex', 'Age', 'SibSp', 'Parch', 'Fare', 'Embarked', 'Title', 'Family_size',
-                     'PersonFare', 'Deck']]
-    return new_data
+from dataframe_manipulations import apply_all
 
 
 def cut_outlier(data):
